+++ CBash/Records/BSGNRecord.cpp
@@ -22,14 +22,14 @@ GPL License and Copyright Notice ============================================
 #include "..\Common.h"
 #include "BSGNRecord.h"
 
-int BSGNRecord::ParseRecord(unsigned char *buffer, const unsigned int &recSize)
+signed long BSGNRecord::ParseRecord(unsigned char *buffer, const unsigned long &recSize)
     {
     if(IsLoaded())
         return -1;
     IsLoaded(true);
-    unsigned int subType = 0;
-    unsigned int subSize = 0;
-    unsigned int curPos = 0;
+    unsigned long subType = 0;
+    unsigned long subSize = 0;
+    unsigned long curPos = 0;
     FormID curFormID = NULL;
     while(curPos < recSize){
         _readBuffer(&subType,buffer,4,curPos);
@@ -61,7 +61,7 @@ int BSGNRecord::ParseRecord(unsigned char *buffer, const unsigned int &recSize)
                 DESC.Read(buffer, subSize, curPos);
                 break;
             case eSPLO:
-                curFormID = new unsigned int;
+                curFormID = new unsigned long;
                 _readBuffer(curFormID,buffer,subSize,curPos);
                 SPLO.push_back(curFormID);
                 break;
@@ -77,12 +77,12 @@ int BSGNRecord::ParseRecord(unsigned char *buffer, const unsigned int &recSize)
     return 0;
     }
 
-unsigned int BSGNRecord::GetSize(bool forceCalc)
+unsigned long BSGNRecord::GetSize(bool forceCalc)
     {
     if(!forceCalc && recData != NULL)
-        return *(unsigned int*)&recData[-16];
-    unsigned int cSize = 0;
-    unsigned int TotSize = 0;
+        return *(unsigned long*)&recData[-16];
+    unsigned long cSize = 0;
+    unsigned long TotSize = 0;
     if(EDID.IsLoaded())
         {
         cSize = EDID.GetSize();
@@ -108,11 +108,11 @@ unsigned int BSGNRecord::GetSize(bool forceCalc)
         TotSize += cSize += 6;
         }
     if(SPLO.size())
-        TotSize += (unsigned int)SPLO.size() * (sizeof(unsigned int) + 6);
+        TotSize += (unsigned long)SPLO.size() * (sizeof(unsigned long) + 6);
     return TotSize;
     }
 
-int BSGNRecord::WriteRecord(_FileHandler &SaveHandler)
+signed long BSGNRecord::WriteRecord(_FileHandler &SaveHandler)
     {
     if(EDID.IsLoaded())
         SaveHandler.writeSubRecord(eEDID, EDID.value, EDID.GetSize());
@@ -123,8 +123,8 @@ int BSGNRecord::WriteRecord(_FileHandler &SaveHandler)
     if(DESC.IsLoaded())
         SaveHandler.writeSubRecord(eDESC, DESC.value, DESC.GetSize());
 
-    for(unsigned int p = 0; p < SPLO.size(); p++)
-        SaveHandler.writeSubRecord(eSPLO, SPLO[p], sizeof(unsigned int));
+    for(unsigned long p = 0; p < SPLO.size(); p++)
+        SaveHandler.writeSubRecord(eSPLO, SPLO[p], sizeof(unsigned long));
     return -1;
     }
 
@@ -133,7 +133,7 @@ void BSGNRecord::Debug(int debugLevel)
     {
     if(!IsLoaded())
         return;
-    unsigned int indentation = 4;
+    unsigned long indentation = 4;
     printf("  BSGN\n");
     if(Header.IsLoaded())
         Header.Debug(debugLevel, indentation);
@@ -153,7 +153,7 @@ void BSGNRecord::Debug(int debugLevel)
         if(debugLevel > 3)
             {
             indentation += 2;
-            for(unsigned int p = 0;p < SPLO.size();p++)
+            for(unsigned long p = 0;p < SPLO.size();p++)
                 {
                 PrintIndent(indentation);
                 printf("%u:%s\n", p, PrintFormID(SPLO[p]));
